backend/api-service/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dnspython==2.0.0; python_version >= '3.6'
email-validator==1.1.2
fastapi==0.62.0
h11==0.11.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2-binary==2.8.6
pydantic==1.7.3; python_version >= '3.6'
python-dotenv==0.15.0
requests==2.25.0
starlette==0.13.6; python_version >= '3.6'
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.12.3
